US8342 Fix linting errors and windows conflicts
